backend/requirements.txt
fastapi
uvicorn
pydantic
fireworks-ai
databases
asyncpg

